Changed all integer to double

+++ Models/DTOs/ProductDTOs/CreateProductRequestModel.cs
@@ -11,8 +11,8 @@ namespace SMS_MVCDTO.Models.DTOs.ProductDTOs
         public string Name { get; set; }
         public string Description { get; set; }
         public double SellingPrice { get; set; }
-        public int Quantity { get; set; }
-        public int ReorderLevel { get; set; }
+        public double Quantity { get; set; }
+        public double ReorderLevel { get; set; }
     }
 
 }
