+++ Models/Entities/Product.cs
@@ -19,8 +19,8 @@ namespace SMS_MVCDTO.Models.Entities
         public string Name { get; set; }
         public string Description { get; set; }
         public double SellingPrice { get; set; }
-        public int Quantity { get; set; }
-        public int ReorderLevel { get; set; }
+        public double Quantity { get; set; }
+        public double ReorderLevel { get; set; }
         public bool IsAvailable { get; set; }
 
     }
